refactor(ui): réecrit le composant Pills en tsx
