+++ packages/ui/src/components/titre.vue
@@ -59,8 +59,7 @@
 <script>
 import Loader from './_ui/loader.vue'
 import { Perimetre } from './_common/perimetre'
-import ActivitesPills from './_common/pills.vue'
-
+import { ActivitesPills } from './activites/activites-pills'
 import { Header as TitreHeader } from './titre/header'
 import { Infos as TitreInfos } from './titre/infos'
 import { Territoires as TitreTerritoires } from './titre/territoires'
