+++ packages/ui/src/components/titres/table-utils.ts
@@ -5,7 +5,7 @@ import { Domaine as CaminoDomaine } from '../_common/domaine'
 import TitreNom from '../_common/titre-nom.vue'
 import TitreTypeTypeNom from '../_common/titre-type-type-nom.vue'
 import CoordonneesIcone from '../_common/coordonnees-icone.vue'
-import ActivitesPills from '../_common/pills.vue'
+import { ActivitesPills } from '../activites/activites-pills'
 import { Statut } from '../_common/statut'
 import { DomaineId } from 'camino-common/src/static/domaines'
 import { TitresTypesTypes } from 'camino-common/src/static/titresTypesTypes'
